third_party/blink/web_tests/platform/generic/http/tests/devtools/compiler-script-mapping-expected.txt

Tests SourceMap and CompilerScriptMapping.


Running: testCompilerScriptMapping
Adding compiled.js
compiled.js UISourceCode arrived
compiled.js script found
source1.js UISourceCode arrived
source2.js UISourceCode arrived
Location checks passed. Requesting content
Content1 arrived.
Content2 arrived.

Running: testInlinedSourceMap
Adding compiled.js
compiled.js UISourceCode arrived
compiled.js script found
source3.js UISourceCode arrived
Location checks passed. Requesting content
<source content> === <source content>

Running: testSourceMapCouldNotBeLoaded
Adding compiled.js
SourceMap Failed to load.
test://evaluations/3/compiler-script-mapping.js
